feat: create Counter common component
